+++ app/src/main/java/com/example/outlaycalc/models/Movement.kt
@@ -3,7 +3,7 @@ package com.example.outlaycalc.models
 import java.util.*
 
 class Movement {
-    var amount: Double? = null
+    var amount: Float? = null
     var description: String = ""
     var outlay: Boolean = true
     var date: Date = Date()
